+++ esti/webhook_server_util.go
@@ -1,36 +1,43 @@
+//nolint:unused
 package esti
+
+// TODO (niro): All the unused errors is because our esti tests filenames are suffixed with _test
+// TODO (niro): WE will need to rename all the esti tests file names to instead using test prefix and not suffix
 
 import (
 	"errors"
 	"fmt"
 	"io"
-	"log"
 	"net"
 	"net/http"
 	"os"
+	"testing"
 	"time"
 )
 
-type hookResponse struct {
-	path        string
-	err         error
-	data        []byte
-	queryParams map[string][]string
-}
-
-type webhookServer struct {
+type WebhookServer struct {
 	s      *http.Server
-	respCh chan hookResponse
+	respCh chan HookResponse
 	port   int
 	host   string
 }
 
-func (s *webhookServer) BaseURL() string {
+const hooksTimeout = 2 * time.Second
+
+var ErrWebhookTimeout = errors.New("timeout passed waiting for hook")
+
+func (s *WebhookServer) BaseURL() string {
 	return fmt.Sprintf("http://%s:%d", s.host, s.port)
 }
 
-func startWebhookServer() (*webhookServer, error) {
-	respCh := make(chan hookResponse, 10)
+func (s *WebhookServer) Server() *http.Server {
+	return s.s
+}
+
+func StartWebhookServer(t testing.TB) *WebhookServer {
+	t.Helper()
+	const channelSize = 10
+	respCh := make(chan HookResponse, channelSize)
 	mux := http.NewServeMux()
 	mux.HandleFunc("/pre-commit", hookHandlerFunc(respCh))
 	mux.HandleFunc("/post-commit", hookHandlerFunc(respCh))
@@ -46,13 +53,13 @@ func startWebhookServer() (*webhookServer, error) {
 	mux.HandleFunc("/post-delete-tag", hookHandlerFunc(respCh))
 	mux.HandleFunc("/timeout", timeoutHandlerFunc(respCh))
 	mux.HandleFunc("/fail", failHandlerFunc(respCh))
-	listener, err := net.Listen("tcp", ":0")
+	listener, err := net.Listen("tcp", ":0") //nolint:gosec
 	if err != nil {
-		return nil, err
+		t.Fatal(err)
 	}
 
 	port := listener.Addr().(*net.TCPAddr).Port
-	fmt.Println("Using port:", port)
+	t.Log("Using port:", port)
 	s := &http.Server{
 		ReadHeaderTimeout: time.Minute,
 		Handler:           mux,
@@ -63,23 +70,24 @@ func startWebhookServer() (*webhookServer, error) {
 	}
 	go func() {
 		if err = s.Serve(listener); err != nil && !errors.Is(err, http.ErrServerClosed) {
-			log.Fatalf("listen: %s\n", err)
+			t.Logf("listen: %s", err)
 		}
 	}()
 
-	return &webhookServer{
+	return &WebhookServer{
 		s:      s,
 		respCh: respCh,
 		port:   port,
 		host:   host,
-	}, nil
+	}
 }
 
-func timeoutHandlerFunc(chan hookResponse) func(http.ResponseWriter, *http.Request) {
+func timeoutHandlerFunc(_ chan HookResponse) func(http.ResponseWriter, *http.Request) {
+	const timeout = 2 * hooksTimeout
 	return func(writer http.ResponseWriter, req *http.Request) {
 		select {
 		case <-req.Context().Done():
-		case <-time.After(2 * hooksTimeout):
+		case <-time.After(timeout):
 		}
 
 		writer.WriteHeader(http.StatusOK)
@@ -87,35 +95,26 @@ func timeoutHandlerFunc(chan hookResponse) func(http.ResponseWriter, *http.Reque
 	}
 }
 
-func failHandlerFunc(chan hookResponse) func(http.ResponseWriter, *http.Request) {
+func failHandlerFunc(_ chan HookResponse) func(http.ResponseWriter, *http.Request) {
 	return func(writer http.ResponseWriter, _ *http.Request) {
 		writer.WriteHeader(http.StatusInternalServerError)
 		_, _ = io.WriteString(writer, "Failed")
 	}
 }
 
-func hookHandlerFunc(respCh chan hookResponse) func(http.ResponseWriter, *http.Request) {
+func hookHandlerFunc(respCh chan HookResponse) func(http.ResponseWriter, *http.Request) {
 	return func(writer http.ResponseWriter, request *http.Request) {
 		data, err := io.ReadAll(request.Body)
 		if err != nil {
-			respCh <- hookResponse{path: request.URL.Path, err: err}
+			respCh <- HookResponse{Path: request.URL.Path, Err: err}
 			_, _ = io.WriteString(writer, "Failed")
 			writer.WriteHeader(http.StatusInternalServerError)
 			return
 		}
 
 		time.Sleep(1 * time.Second) // Added sleep to differentiate between event timestamps
-		respCh <- hookResponse{path: request.URL.Path, data: data, queryParams: request.URL.Query()}
+		respCh <- HookResponse{Path: request.URL.Path, Data: data, QueryParams: request.URL.Query()}
 		_, _ = io.WriteString(writer, "OK")
 		writer.WriteHeader(http.StatusOK)
-	}
-}
-
-func responseWithTimeout(s *webhookServer, timeout time.Duration) (*hookResponse, error) {
-	select {
-	case res := <-s.respCh:
-		return &res, nil
-	case <-time.After(timeout):
-		return nil, errors.New("timeout passed waiting for hook")
 	}
 }